[fastmanifest] fix crash in hybridmanifest.__nonzero__()
Summary:
hybridmanifest.__nonzero__() was explicitly trying to call __nonzero__() on the
underlying manifest object.  This breaks when the underlying manifest is a
manifestdict.  manifestdict does not implement __nonzero__ (it instead has a
__len__ method that gets used instead when evaluating the manifestdict as a
boolean).

Because of this issue I was getting a crash in a local script when calling
repo.commitctx() with a memctx object.

Test Plan: Confirmed that calling repo.commitctx() with a memctx no longer crashes.

Reviewers: #sourcecontrol, durham, quark

Reviewed By: quark

Subscribers: net-systems-diffs@, yogeshwer, mjpieters

Differential Revision: https://phabricator.intern.facebook.com/D4191858

Signature: t1:4191858:1479330743:7adb668c9529a8c8abaf9a8a1e64c8ee78b64d86