treemanifest: catch pyexception for __nonzero__
Summary:
The recently added __nonzero__ function was not handling the possibility of
pyexceptions. This adds the appropriate catch.

A pyexception is where somewhere up the stack something has set the python error
string, and we just need to return the appropriate error value from the top
level c api to indicate an error happened.

Test Plan: none. I just caught this while debugging

Reviewers: #mercurial

Differential Revision: https://phabricator.intern.facebook.com/D4202090