lock: make sync check more robust
Summary:
Previously we detected if we were in the middle of a sync by checking if the
sync limiter lock was taken. This lock is not actually taken during a sync where
we force it to sync. So let's replace _synclock with a more robust
self._issyncing boolean.

Test Plan:
Did some manual testing with break points. My last attempt was just
naive, so this isn't a big change.

Reviewers: #mercurial, quark

Reviewed By: quark

Differential Revision: https://phabricator.intern.facebook.com/D4289240

Signature: t1:4289240:1481071402:e2a448a2739bff3eecfea90a6fd7a0376b944adb