lock: log lock time for failures
Summary:
Previously we were not logging sqllock wait times for waits that timed out. This
adds that, and adds a variable to indicate if the lock succeed or failed.

Test Plan: Added a test for the ui.log metrics

Reviewers: #mercurial, quark

Reviewed By: quark

Subscribers: quark

Differential Revision: https://phabricator.intern.facebook.com/D4295786

Signature: t1:4295786:1481154111:442b55f230b53add4e0966223af165f5195ba8b4